Add podSubnet to templates
