+++ pkg/minikube/bootstrapper/bsutil/template/template.go
@@ -19,137 +19,7 @@ package template
 import (
 	"fmt"
 	"sort"
-	"text/template"
 )
-
-// KubeAdmConfigTmplV1Alpha1 is for Kubernetes v1.11
-var KubeAdmConfigTmplV1Alpha1 = template.Must(template.New("configTmpl-v1alpha1").Funcs(template.FuncMap{
-	"printMapInOrder": printMapInOrder,
-}).Parse(`apiVersion: kubeadm.k8s.io/v1alpha1
-kind: MasterConfiguration
-{{if .NoTaintMaster}}noTaintMaster: true{{end}}
-api:
-  advertiseAddress: {{.AdvertiseAddress}}
-  bindPort: {{.APIServerPort}}
-  controlPlaneEndpoint: localhost
-kubernetesVersion: {{.KubernetesVersion}}
-certificatesDir: {{.CertDir}}
-networking:
-  serviceSubnet: {{.ServiceCIDR}}
-etcd:
-  dataDir: {{.EtcdDataDir}}
-nodeName: {{.NodeName}}
-{{if .ImageRepository}}imageRepository: {{.ImageRepository}}
-{{end}}{{if .CRISocket}}criSocket: {{.CRISocket}}
-{{end}}{{range .ExtraArgs}}{{.Component}}ExtraArgs:{{range $i, $val := printMapInOrder .Options ": " }}
-  {{$val}}{{end}}
-{{end}}{{if .FeatureArgs}}featureGates: {{range $i, $val := .FeatureArgs}}
-  {{$i}}: {{$val}}{{end}}
-{{end}}`))
-
-// KubeAdmConfigTmplV1Alpha3 is for Kubernetes v1.12
-var KubeAdmConfigTmplV1Alpha3 = template.Must(template.New("configTmpl-v1alpha3").Funcs(template.FuncMap{
-	"printMapInOrder": printMapInOrder,
-}).Parse(`apiVersion: kubeadm.k8s.io/v1alpha3
-kind: InitConfiguration
-apiEndpoint:
-  advertiseAddress: {{.AdvertiseAddress}}
-  bindPort: {{.APIServerPort}}
-bootstrapTokens:
-  - groups:
-      - system:bootstrappers:kubeadm:default-node-token
-    ttl: 24h0m0s
-    usages:
-      - signing
-      - authentication
-nodeRegistration:
-  criSocket: {{if .CRISocket}}{{.CRISocket}}{{else}}/var/run/dockershim.sock{{end}}
-  name: {{.NodeName}}
-  taints: []
----
-apiVersion: kubeadm.k8s.io/v1alpha3
-kind: ClusterConfiguration
-{{if .ImageRepository}}imageRepository: {{.ImageRepository}}
-{{end}}{{range .ExtraArgs}}{{.Component}}ExtraArgs:{{range $i, $val := printMapInOrder .Options ": " }}
-  {{$val}}{{end}}
-{{end -}}
-{{if .FeatureArgs}}featureGates: {{range $i, $val := .FeatureArgs}}
-  {{$i}}: {{$val}}{{end}}
-{{end -}}
-certificatesDir: {{.CertDir}}
-clusterName: kubernetes
-controlPlaneEndpoint: localhost:{{.APIServerPort}}
-etcd:
-  local:
-    dataDir: {{.EtcdDataDir}}
-kubernetesVersion: {{.KubernetesVersion}}
-networking:
-  dnsDomain: {{if .DNSDomain}}{{.DNSDomain}}{{else}}cluster.local{{end}}
-  podSubnet: {{if .PodSubnet}}{{.PodSubnet}}{{else}}""{{end}}
-  serviceSubnet: {{.ServiceCIDR}}
----
-apiVersion: kubelet.config.k8s.io/v1beta1
-kind: KubeletConfiguration
-evictionHard:
-  nodefs.available: "0%"
-  nodefs.inodesFree: "0%"
-  imagefs.available: "0%"
-`))
-
-// KubeAdmConfigTmplV1Beta1 is for Kubernetes v1.13+
-var KubeAdmConfigTmplV1Beta1 = template.Must(template.New("configTmpl-v1beta1").Funcs(template.FuncMap{
-	"printMapInOrder": printMapInOrder,
-}).Parse(`apiVersion: kubeadm.k8s.io/v1beta1
-kind: InitConfiguration
-localAPIEndpoint:
-  advertiseAddress: {{.AdvertiseAddress}}
-  bindPort: {{.APIServerPort}}
-bootstrapTokens:
-  - groups:
-      - system:bootstrappers:kubeadm:default-node-token
-    ttl: 24h0m0s
-    usages:
-      - signing
-      - authentication
-nodeRegistration:
-  criSocket: {{if .CRISocket}}{{.CRISocket}}{{else}}/var/run/dockershim.sock{{end}}
-  name: {{.NodeName}}
-  taints: []
----
-apiVersion: kubeadm.k8s.io/v1beta1
-kind: ClusterConfiguration
-{{ if .ImageRepository}}imageRepository: {{.ImageRepository}}
-{{end}}{{range .ExtraArgs}}{{.Component}}:
-  extraArgs:
-{{- range $i, $val := printMapInOrder .Options ": " }}
-    {{$val}}
-{{- end}}
-{{end -}}
-{{if .FeatureArgs}}featureGates:
-{{range $i, $val := .FeatureArgs}}{{$i}}: {{$val}}
-{{end -}}{{end -}}
-certificatesDir: {{.CertDir}}
-clusterName: kubernetes
-controlPlaneEndpoint: localhost:{{.APIServerPort}}
-dns:
-  type: CoreDNS
-etcd:
-  local:
-    dataDir: {{.EtcdDataDir}}
-kubernetesVersion: {{.KubernetesVersion}}
-networking:
-  dnsDomain: {{if .DNSDomain}}{{.DNSDomain}}{{else}}cluster.local{{end}}
-  podSubnet: ""
-  serviceSubnet: {{.ServiceCIDR}}
----
-apiVersion: kubelet.config.k8s.io/v1beta1
-kind: KubeletConfiguration
-imageGCHighThresholdPercent: 100
-evictionHard:
-  nodefs.available: "0%"
-  nodefs.inodesFree: "0%"
-  imagefs.available: "0%"
-`))
 
 // printMapInOrder sorts the keys and prints the map in order, combining key
 // value pairs with the separator character
