+++ pkg/util/constants.go
@@ -24,9 +24,8 @@ import (
 
 // These constants are used by both minikube
 const (
-	APIServerPort      = 8443
-	DefaultDNSDomain   = "cluster.local"
-	DefaultServiceCIDR = "10.96.0.0/12"
+	APIServerPort    = 8443
+	DefaultDNSDomain = "cluster.local"
 )
 
 // DefaultV114AdmissionControllers are admission controllers we default to in v1.14.x
